+++ src/enzo/rk4.C
@@ -38,9 +38,9 @@ c
 
 #define IDX(a,b,c) ( ((c)*jn + (b))*in + (a) )
 
-int rk4(double *repsi, double *impsi,
+int rk4(float *repsi, float *impsi,
          int in, int jn, int kn, int rank,
-         double dt, double dx[], double dy[], double dz[],
+         float dt, float dx[], float dy[], float dz[],
          double hmcoef)
 
 {
@@ -51,14 +51,14 @@ int rk4(double *repsi, double *impsi,
   int i, j, k, jsm1, ksm1, jep1, kep1, ism2, jsm2, ksm2, jep2, kep2;
   int size = in*jn*kn;
 
-  double *r1 = new double[size];
-  double *i1 = new double[size];
-  double *r2 = new double[size];
-  double *i2 = new double[size];
-  double *r3 = new double[size];
-  double *i3 = new double[size];
-  double *r4 = new double[size];
-  double *i4 = new double[size];
+  float *r1 = new float[size];
+  float *i1 = new float[size];
+  float *r2 = new float[size];
+  float *i2 = new float[size];
+  float *r3 = new float[size];
+  float *i3 = new float[size];
+  float *r4 = new float[size];
+  float *i4 = new float[size];
 
   /* ======================================================================= */
